requirement.txt
streamlit
tensorflow
opencv-python
pillow
numpy
gdown

